New Gun Config System is Live (#636)
* Bugfixes and improvements.

* Fixed xenos not dying in a burning inferno. The xeno life procs are in general a mess and need to be rewritten, so this is mostly a patch. If xenos are not seeing the death messages properly, that means they are not added to the aliens list at round start/during the round. For example, anyone transformed into a xeno via the admin verb won't see death messages. This goes for all xeno_message() things.

* Preparing for a merge.

* Committing to merge with Joshuu's update.

* Hotfix for hot predator spawning action.

* Hotfix for opening shutters in flight.

* Hotfix for poddoors.

* Committing some work in progress changes.

* Small changes, pending more work.

* Develop/hotfix cm (#1)

* Hotfixes for preds and poddoors.

* Committing monkey's updated helmet garb dmi.

* Fixed a few runtimes.
Disabled gun restrictions on the WO outpost map.
Added debug toggles for easy debug messages for a few things.
Added /image support to the recycler. Should now handle images instead of deleting them, when needed. This also makes projectiles more optimized as the muzzle flash/bounce is recycled. It's also possible to recycle overlays and should probably be done as such.
Cleaned up some gun code to get rid of some useless procs I left in as placeholder and corrected a bunch of incorrect bit flag checks.
Optimized update icon on guns so they're not constantly recycling overlays for no reason whatsoever.
Reorganized the predator items file with some minor changes. Also added styles to the various messages.
Removed healing internal damage with the crystal by using it in combination with a knife. It's already plenty healing.
Reorganized held items along with some other icon name changes. item_state will now be used for back icons on the sprite itself, instead of icon_state as the default. This is consistent with the other slots.
IMPORTANT NOTE ON GUNS::: ICON STATE and ITEM STATE should match each other, so we can easily identify them. Gun icons are named like so: m41a. All letters are lowercase. If it has a unique skin, it should be prefaced with a letter like s_m41a (snow camo). If it's a unique item (not just a skin), add a letter/number to the front of the name, like m41a2. The empty state of the gun should follow ICON STATE with a _e, like m41a_e. Wielded items in ITEM STATE are followed by _w, like m41a_w.
Converted gun attachment lists to base off the gun instead of the attachment.
Rewrote how guns calculate their damage and stuff. It's all done through combat_defines in config, which is then passed on to config in the game. There are also some defaults in case the file is missing.
Full list of changes in the changelog.

* Made two more weapons unacidable.

* Fixed a few runtimes.
Disabled gun restrictions on the WO outpost map.
Added debug toggles for easy debug messages for a few things.
Added /image support to the recycler. Should now handle images instead of deleting them, when needed. This also makes projectiles more optimized as the muzzle flash/bounce is recycled. It's also possible to recycle overlays and should probably be done as such.
Cleaned up some gun code to get rid of some useless procs I left in as placeholder and corrected a bunch of incorrect bit flag checks.
Optimized update icon on guns so they're not constantly recycling overlays for no reason whatsoever.
Reorganized the predator items file with some minor changes. Also added styles to the various messages.
Removed healing internal damage with the crystal by using it in combination with a knife. It's already plenty healing.
Reorganized held items along with some other icon name changes. item_state will now be used for back icons on the sprite itself, instead of icon_state as the default. This is consistent with the other slots.
IMPORTANT NOTE ON GUNS::: ICON STATE and ITEM STATE should match each other, so we can easily identify them. Gun icons are named like so: m41a. All letters are lowercase. If it has a unique skin, it should be prefaced with a letter like s_m41a (snow camo). If it's a unique item (not just a skin), add a letter/number to the front of the name, like m41a2. The empty state of the gun should follow ICON STATE with a _e, like m41a_e. Wielded items in ITEM STATE are followed by _w, like m41a_w.
Converted gun attachment lists to base off the gun instead of the attachment.
Rewrote how guns calculate their damage and stuff. It's all done through combat_defines in config, which is then passed on to config in the game. There are also some defaults in case the file is missing.
Full list of changes in the changelog.