Fixes pipeline and pipe_network datums being incorrectly initiated because the the new process scheduler was calling the atmos controller setup() before the object controller setup (the opposite of the old master controller system). For same reason I moved the ticker controller file to be at the top. Also fixed a typo in the sun controller.
Note that with this change the air_system controller setup is after the objects controller (which is the opposite of what the old master controller did) but I didn't see any runtime occuring. If something bad ever happens with that, you can move the air_system setup into the objects controller setup (before the object initialization part), which is confusing but was the old way.