Fixes predator dropping some red gibs on gib(). Fixes #3356
Rewrote how we handle the gib animation, we no longer make the mob and its gib proc wait for the gib animation to finish to delete the mob. We now spawn a temp overlay object that handles the gib animation and we cdel() the mob right away. Same thing for dust().

Properly typecasts the human var/shoes and var/gloves, so we don't have to check whether our gloves are...gloves.

Fixes sprite issue on the "Business Person" in the Select Equipment options.

Xeno queen is gibbable again.

The message shown when a human's limb is ripped now only appears for people close (3 tiles), this should reduce the limb drop message spam when the Alamo crashlands on a few marines.