Initial commit for feature.
Based on #4587.

The intent of this change is to allow the quick-equip built-in hotkey E to put weapons, tools, ammo, and other items into equipped containers in a rational order.

Select-equipment and other mostly-internal uses of the procs edited here should remain identically functional.

In this commit, pushing E will attempt to store the item in the active hand to body slots/containers in the following order:

1. Back
2. ID
3. Uniform
4. Armor/Jacket Body Slot
5. Mask
6. Helmet
7. Boots
8. Gloves
9. Headset
10. Eyewear
11. Belt Item's Storage (ammo belt, toolbelt, etc. in belt slot)
12. Back Scabbard
13. Belt Slot (belts themselves)
14. Armor Slot Item's Storage (gunbelts/scabbards placed here)
15. Left Pouch Storage
16. Right Pouch Storage
17. Armor Storage Slot (where mag-harn'd guns go)
18. Left Pocket (the slot itself)
19. Right Pocket (the slot itself)
20. Backpack Storage