Hydroponics n stuff
changes:
  - bugfix: "Fixed shutters on Big Red and Almayer research being the wrong orientation."
  - bugfix: "Fixed the M94 marking flare packs not containing 8 signal flares. #6797"
  - experiment: "Research now has a hydroponics laboratory."
  - rscadd: "Added a Hydroponics Research Locker that contains the necessary items, including three spray bottles of ammonia. (That you can empty and fill with better mixes)."
  - rscadd: "Added a disk and a seed cabinet to store flora disks and plant seeds."
  - rscadd: "Added Dinitroaniline, a herbicide and pesticide mixture for use in hydroponics. Can be made by reacting ammonia, nitrogen and sulfuric acid."
  - tweak: "Flora disk boxes now contain 14 disks instead of 7."
  - tweak: "Harmful plants will no longer randomly appear in weeded hydroponics trays."
  - tweak: "Hydroponics trays now have a flush function, you can use this to reset the tray."
  - tweak: "Tweaked how reagents react with hydroponics."
  - tweak: "Different chemicals can be added to hydroponics trays. They can act as nutrient, toxin, anti-toxin, weedkiller, pestkiller, mutator or as a yield modifier."
  - tweak: "Sprays can now add reagents to glass containers and hydroponics trays. Make your own custom weedkiller or plant nutrient spray!"

This should add some more for researchers to do, and the laboratory will be useful in the planned future chem updates.

I made sure that harmful or hallucinogenic plants doesn't randomly spawn in the trays. And they are also not accessible from the seed vendor unless contraband is accessed. Plasticide has been removed from the vendor completely, so researchers can't powergame and make plasteel.

Removed the furthest west dropship crash location on the lower deck, as this have caused a round with xenos being trapped resulting in a terrible stalemate. Compensated by adding a new crash location in brig. Didn't include this in the changelog as we don't disclose dropship crashsites.

Closes: #6797 #6807