Torii fixed, can no longer insert hot items into storage.
This includes lit flares, burning matches, lit welding tools, lit lighters etc. Technically also the plasma cutter, but that does nothing anymore and only really exists for map flavour.

While the torii arch could be message spammed with a flame source, it could not actually be lit. It was also banged on during lighting attempts or just with objects in general, which seems disrespectful.