Sorokyne runtime errors fix (#1645)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Fixes #1629 - turns out the first two rows of Sorokyne were empty which
caused a lot of runtime errors (like 300 each map load) (thanks Fira for
figuring it out):


![image](https://user-images.githubusercontent.com/964559/202930777-e71fbc2c-0062-435e-bfba-45db953a659c.png)

So I filled the first two rows with more of what was above them, so the
tunnel got a bit extended:


![image](https://user-images.githubusercontent.com/964559/202930696-ee7cc9db-4f03-462e-976e-93db47e0e70b.png)

A good deal of space was added, and the right side was a bit extended as
well:


![image](https://user-images.githubusercontent.com/964559/202930723-1a8281ae-7dfc-4f53-8ac7-2513c6e88ee8.png)

Now it loads without an issue.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Runtime errors caused unexpected things not to load on the map, plus
bloat the logs. Best not to let that happen.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

## Changelog

<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
fix: Sorokyne Strata should no longer experience a lot of runtime errors
caused by empty tiles being present in the map.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->