Small fixes and tweaks (#1646)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Saw a few minor bugs and things that could be improved, so I fixed /
tweaked them. Read the changelog.

For the Working Joe stuff - when I played them, it always took me a
second or two to remember that "alwaysknow" is the way to invoke "the
Working Joe line", so I added a secondary way to trigger that in a way
that I associated it - `*workingjoe`. Same with the mess line. But
apparently the parser likes to remove the final "s", so I added a backup
way of invoking it, `*mes`.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

## Changelog

It improves the experience of the game to make the errors less
noticeable.

<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
fix: Mine boxes can now hold all the mines they are supposed to hold.
fix: Hydroponics locker now starts with the correct lock indicator.
spellcheck: Ravager's abilities now have proper capitalization.
qol: Some Working Joe emotes have an alternative way of invoking them
now to help you embody the character easier.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->