Changes languages to German and Spanish. (#1681)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Renames German and Spanish to be simply that.

'Weltraumdeutsch' -> 'German'
'Neospanish' -> 'Spanish'

## Why It's Good For The Game

All the other languages follow this naming scheme, so it's pretty odd
that these two in particular get the special snowflake treatment. Also,
the lore monkeys wanted it.


![image](https://user-images.githubusercontent.com/82319489/203631027-4cb32078-6883-4d56-b9bb-df48d7a5acbc.png)


## Changelog

<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
spellcheck: Renamed German and Spanish.
/:cl:

(There's no tweak label for the CL, apparently?)

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->

Co-authored-by: F.F <davidkramer1804@gmail.com>