Changes the PR template. (#1698)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# Document the changes in your pull request.

Revises PR template to adding a testing and issue closer section.

Maintainability starts with the PR.

# Explain why it's good for the game.

Not user facing.

QOL for longer term DevOps. When dealing with unexpected behaviors in
the wild, knowing if it is intended can be helpful in some cases.
Moreover, it provides a place for testing/images to be present without
getting lost in long lines of commits/comments/etc.

This would also allow for a seamless way to keep a one stop place for
documentation in cases of larger PRs.

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<details>
<summary>Screenshots&Videos</summary>


![image](https://user-images.githubusercontent.com/82319489/204013999-6308edc1-638a-456c-ae16-d17dc5712603.png)

</details>


</details>


# Changelog

<!-- Edit the changelog below to reflect the changes made by this PR,
even if the changes are minor - required for every PR.
If you add a name after the ':cl', that name will be used in the
changelog. Leave it empty to use your GitHub name. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Maintainers freely
reserve the right to remove and add tags should they deem it
appropriate. -->



<!-- Both :cl:'s are required for the changelog to work! -->

Co-authored-by: F.F <davidkramer1804@gmail.com>
Co-authored-by: harryob <me@harryob.live>