Refilling a Pressurized Reagent Canister Pouch refills its autoinjector (#1805)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

It's a touch annoying when you refill a Pressurized Reagent Canister
Pouch you have to then take out its autoinjector and put it back in for
it to be filled, and then you have to top up the Pressurized Reagent
Canister Pouch again. So I made it that you only need to refill the
Pressurized Reagent Canister Pouch to fill both it and its autoinjector
when you use the Reagent Tank.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

A simple QoL to make Pressurized Reagent Canister Pouches less fiddly to
deal with.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

## Changelog

<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
qol: When you refill a Pressurized Reagent Canister Pouch with a Reagent
Tank you now also automatically refill its autoinjector so you will be
fully topped up in one click!
code: Simplified a casting of an autoinjector to a too detailed type
during Pressurized Reagent Canister Pouch initialisation.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->