Simple Walking Implementation (#1597)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Implements the UI button and keybinding to toggle movement intents, as
well as toggling to walk and slowing down. Also supports xenos using
walking icon_states, but only if they have walking states in the first
place.

Ideally this acts as a foundation to implement more whacky movement
intent based mechanics, like slipping on wet floors if you're running,
or xenos gaining invulnerability to all threats, physical or magical.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

It's neat, I guess. A relic of the old times that can act as a fun
gameplay mechanic, especially if the tactics get expanded upon in the
future. If not, extreme roleplayers can just have fun, walking around.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

## Changelog

<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
add: Re-added the basics of walking.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>