Warrior Nerf  {no gpb update} (#1798)
## About The Pull Request

This PR nerfs warrior. I believe it is a little bit too powerful
compared to when it had shields. The rework intended it the caste to fit
in more as a side/frontline ambusher however it allowed the caste to get
away with very nasty cooldown reductions due to the inherently low
cooldown of Fling and Punch. I've also reduced the amount of CDR gained
as a way to curb the issue with current warrior.

## Why It's Good For The Game

Fine tuning, has been left alone for far too long. Although current
warrior is very fun and actually fits in with its role, the rework did
not take into account the cooldowns of fling/punch which is far too
small for the caste along with the slash CDR. This PR will fix it, and
hopefully stop people from whining.

## Changelog usnpeepoo
:cl:
balance: Warriors slash cooldown reduction has been reduced by 0.10
seconds
balance: Warriors Fling, and Punch have their cooldowns increased by 1
second
/:cl: