Chances Changes: “I say we take off and nuke the entire site from orbit. It’s the only way to be sure.” (#1785)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR fixes and changes some areas in LV522 Chances Claim

## Why It's Good For The Game

[I said so](https://www.youtube.com/watch?v=rniukPhgptU&t=1923s)

## Changelog

<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl: SpartanBobby
balance: A-Block Droms north-south hallways are now 1 tile wider on each
side with a 2x3 destructible blocker in the middle to hinder snipers and
boilers
balance: A-Block Drom south hallway now has a door leading outside
balance: Some of the man-made barricades on the map have now been
destroyed prior to the Almayer landing
balance: LV522 Chances Claim: Removed a lot of 1x1 tunnels and redundant
areas in the reactor
fix: fixed a strange bug with unbreakable cargo containers
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->