Gave SLs other kits to start with (#1464)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

I've seen a number of times the SLs just take their starting pyro kit
and toss it out for someone in their squad to take, which isn't a great
thing if a large chunk of your essential kit is not used. So I figured -
why not give the SLs some variety of what they can pick up to define
their kit.

So the kits are mostly the kits you can find in Requisitions, plus the
mini-pyro kit they had, an L42A kit if they want to go for something
with IFF, and a basic engineering supplies kit which has a full
electronics pouch (3 batteries, 3 apc boards), a construction pouch with
50 metal, 30 plasteel and 15 barbed wire, and welding goggles, so they
only need to grab a toolbelt and they can fully repair the comms by
themselves and do some basic defences around it.


![image](https://user-images.githubusercontent.com/964559/203536145-ed9535bb-3020-479f-b594-e067a63a95c9.png)

Hopefully these will appeal to most SLs without being too unbalanced. If
you think the numbers in the engineering kit are off or some of the kits
don't fit well with the role, let me know, I'm a bit out of touch with
the current meta.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More gear options means the players will be able to have the experience
they want without having to bother Requisitions.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

## Changelog

<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
add: SLs can now pick from a variety of starting weaponry kits rather
than be forced to get the pyro kit.
code: Small refactor of flags around which items you can vend from a
vendor.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->