Whiskey Outpost fixes and rank changes (#1687)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This adds to a small PR of mine and makes it into a larger one which
contains the following: spelling fixes, changes to two job descriptions
and a minor shuffle in the ranks of four jobs (Ground Commander,
Lieutenant Commander, Quartermaster and Mortar Crew).
Fixes: #1682
removes the welding goggles from the Squad Engineers, as they already
have welding helmets
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
### Spelling fixes
Always good to have
### description changes
Changes a reference to the Lieutenant Command who was called a "Gunnery
Sergeant" when that is not his current rank (Captain) or the rank that I
am changing him to (Second Lieutenant). He is just now referenced as his
job title, Lieutenant Commander
### Rank changes
As I've explained in a discord post, the CO having the rank of Major
while Captain Naiche, the canonical Commander of all marine forces on
LV, is giving them orders.
This a leftover from the rank rework when the Ground Commander just had
the rank of Commander, which put him one rank below the Captain.
Dropping the CO to the difference from before, which in this case is
First Lieutenant remedies this lore conflict.
Entailing this also forces me to drop the XO (who was a captain) to
Second Lieutenant and the QM (who was a First Lieutenant) to First
Sergeant. I don't think these changes are a detriment to the game as all
marines besides the CO, XO and QM were enlisted men anyhow, which means
it doesn't really matter if the CO is a Captain or First Lieutenant. The
QM is still the third highest ranking marine, just no longer an officer.
I am also elevating the Mortar Crew from PFC to Lance Corporal as I
think a man that went through specialized training which is now his
dedicated assignment ought to outrank a rifleman, like a squad engie or
medic.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding, and may discourage maintainers from reviewing or merging
your PR. This section is not strictly required for (non-controversial)
fix PRs or backend PRs. -->

## Changelog
:cl: Googles_Hands
fix: Changed the WO CO, XO, QM and Mortar Crew ranks to be more lore
friendly
fix: Changed the WO reinforcement announcement to be more lore friendly
fix: Fixed the description of the WO XO and CMP in reference to the XO's
title
spellcheck: Fixed a few misspellings in WO announcements
fix: removes redundant welding protection from squad engies
/:cl: